moved mongo connection and promise to db.js
